pyproject.toml
[tool.poetry]
name = "tinamit"
version = "3.0.0"
description = "Conexión de modelos socioeconómicos (dinámicas de los sistemas) con modelos biofísicos."
authors = ["julienmalard <julien.malard@mail.mcgill.ca>"]
license = "AGPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.8"
trio = "^0.17.0"
click = "^7.1.2"
numpy = "^1.19.2"
xarray = "^0.16.1"
ennikkai = "^1.2.5"
pysd = "^0.10.0"
pandas = "^1.1.2"
scipy = "^1.5.2"
regex = "^2020.7.14"
python-dateutil = "^2.8.1"
asks = "^2.4.10"
tldextract = "^2.2.3"
semantic-version = "^2.8.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
